Oops, broken build.

+++ src/cpucore/cpucore.go
@@ -105,7 +105,7 @@ func (c *Core) ClockIn() {
 		c.alu.ReadEval()
 	}
 	if c.getDecoderFlag(instruction.AccInst) >= 0 {
-		c.alu.ExectuteAccInst(c.getDecoderFlag(instruction.AccInst))
+		c.alu.ExectuteAccInst(uint64(c.getDecoderFlag(instruction.AccInst)))
 	}
 
 	// Finally, any internal bus loads. Do this last to make sure the bus has valid data
